# proviemplea/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea",
        "description": "Documentación oficial de la API ProviEmplea.\n\n**Rate Limiting:** Máximo 60 solicitudes por minuto por IP.\n**Caché:** Los endpoints GET tienen caché de 5 minutos.\n**Tiempos de respuesta esperados:** GET ~100ms | POST ~200ms | PUT ~150ms | DELETE ~100ms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas activas",
                "description": "Retorna el listado de todas las empresas con estado activo.",
                "operationId": "9239adbd1cfe29bdf0372db2124d0b1f",
                "parameters": [
                    {
                        "name": "rubro",
                        "in": "query",
                        "description": "Filtrar por rubro: Tecnología, Salud, Educación",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Tecnología"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Empresa"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "661f9511-f39c-42d5-b817-557766551111",
                                        "nombre_empresa": "NTT DATA Chile",
                                        "rut_empresa": "76.123.456-7",
                                        "email": "contacto@nttdata.cl",
                                        "tipo_empresa": "Privada",
                                        "rubro": "Tecnología",
                                        "beneficios": [
                                            "Seguro de salud",
                                            "Home office",
                                            "Capacitaciones"
                                        ],
                                        "validado": true,
                                        "activo": true
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Crear nueva empresa",
                "description": "Registra una nueva empresa. El RUT debe ser único.",
                "operationId": "bea7e1cd70745d9a85563a4cdba64902",
                "requestBody": {
                    "description": "Datos de la empresa a registrar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Empresa"
                            },
                            "example": {
                                "nombre_empresa": "Startup Verde SpA",
                                "rut_empresa": "77.987.654-3",
                                "email": "hola@startupverde.cl",
                                "tipo_empresa": "Startup",
                                "rubro": "Sustentabilidad",
                                "beneficios": [
                                    "Flexibilidad horaria",
                                    "Trabajo remoto",
                                    "Bono anual"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación — RUT duplicado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "description": "Retorna los datos de una empresa usando su UUID.",
                "operationId": "7e0d938df2a2305cd42451cc84ec72ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "661f9511-f39c-42d5-b817-557766551111"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa completa (PUT)",
                "description": "Actualiza todos los datos de una empresa registrada.",
                "operationId": "61d6235f805bd62afcf89b24a6bae3aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "661f9511-f39c-42d5-b817-557766551111"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Empresa"
                            },
                            "example": {
                                "nombre_empresa": "NTT DATA Chile Actualizado",
                                "rut_empresa": "76.123.456-7",
                                "email": "nuevo@nttdata.cl",
                                "tipo_empresa": "Privada",
                                "rubro": "Tecnología e Innovación",
                                "beneficios": [
                                    "Seguro de salud",
                                    "Home office",
                                    "Capacitaciones",
                                    "Bono anual"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Eliminar empresa",
                "description": "Borrado lógico: pone activo=false. La empresa no aparece en listados pero sus datos se conservan.",
                "operationId": "ba97e99dc7744b02f562b861b32e29a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "661f9511-f39c-42d5-b817-557766551111"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Empresa desactivada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/empresas/{id}/toggle": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Activar o desactivar empresa (PATCH)",
                "description": "Cambia el estado activo/inactivo de la empresa. Operación de encender o apagar.",
                "operationId": "eee8b6b8c1104d36506531c794625f90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "661f9511-f39c-42d5-b817-557766551111"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado cambiado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estado de empresa actualizado correctamente"
                                        },
                                        "activo": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empresa": {
                "required": [
                    "nombre_empresa",
                    "rut_empresa"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "661f9511-f39c-42d5-b817-557766551111"
                    },
                    "nombre_empresa": {
                        "type": "string",
                        "example": "NTT DATA Chile"
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76.123.456-7"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "contacto@nttdata.cl"
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "example": "Privada"
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Tecnología"
                    },
                    "beneficios": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Seguro de salud"
                        }
                    },
                    "validado": {
                        "type": "boolean",
                        "example": false
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}